N-dimensional refactoring
The SCIFIO Metadata model is now N-dimensional, instead of the fixed 5-D
as implemented by Bio-Formats.

This allows an arbitrary number of axes per image. Ordering of axes is
still indicated by the order in the metadata, but valid orders are no
longer restricted to "XY....".

A new attribute, "planar dimensions", indicates the number of dimensions
that define a single image plane. For example, if given an image with
X=256, Y=256, C=3 and 2 planar dimensions, the image is 3 256x256
planes. If there are 3 planar dimensions for the same dataset, then the
C is the RGB channel. If there were 3 planar dimensions and the order
was CXY instead, it would indicate an RGB interleaved dataset.

There is a semi-universal assumption that the X and Y axes are still
present as planar axes. Removal of this assumption would require an
additional refactoring.

This single commit makes all Metadata changes and updates all utility
classes and Formats to comply to the new N-D model.

+++ scifio/src/main/java/io/scif/img/cell/SCIFIOCellImgFactory.java
@@ -245,8 +245,8 @@ public final class SCIFIOCellImgFactory<T extends NativeType<T>> extends
 	 *         [1];
 	 */
 	public static int[] getOptimalCellXY(final Reader reader) {
-		return getOptimalCellXY(reader, reader.getOptimalTileWidth(0), reader
-			.getOptimalTileHeight(0));
+		return getOptimalCellXY(reader, (int) reader.getOptimalTileWidth(0),
+			(int) reader.getOptimalTileHeight(0));
 	}
 
 	/**
@@ -272,8 +272,8 @@ public final class SCIFIOCellImgFactory<T extends NativeType<T>> extends
 
 		final Metadata meta = reader.getMetadata();
 
-		final int sizeX = meta.getAxisLength(0, Axes.X);
-		final int sizeY = meta.getAxisLength(0, Axes.Y);
+		final int sizeX = (int)meta.getAxisLength(0, Axes.X);
+		final int sizeY = (int)meta.getAxisLength(0, Axes.Y);
 
 		// Invalid sizes default to 1, 1, which are automatically expanded.
 		if (tileWidth <= 0) tileWidth = 1;
